Organise context implementation into test path
