+++ composer.json
@@ -12,7 +12,8 @@
     "autoload-dev": {
         "psr-4": {
             "Php\\PieIntegrationTest\\": "test/integration/",
-            "Php\\PieUnitTest\\": "test/unit/"
+            "Php\\PieUnitTest\\": "test/unit/",
+            "Php\\PieBehaviourTest\\": "test/behaviour/"
         }
     },
     "authors": [
